# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAPhoto API
  description: |-
    Keep in touch with your friends by sharing photos of special moments, thanks to WASAPhoto! You can
    upload your photos directly from your PC, and they will be visible to everyone following you.
  version: "3.0"
...
paths:
...
  /users/{user_id}/followers:
    get:
      tags: ["followers"]
      summary: Gets user's followers
      description: Get the followers list of the user
      operationId: getUserFollowers
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          schema:
            $ref: "#/components/schemas/uid"
          required: true
          description: The user ID of the user to follow.
      responses:
        '200':
          description: Returns the user list
          content:
            application/json:
              schema:
                type: array
                description: The list of followers.
                minItems: 0
                maxItems: 100
                items:
                  $ref: "#/components/schemas/uid_name"
              example:
                - user_id: "123e4567-e89b-12d3-a456-426655440000"
                  username: "Maria"
                - user_id: "123e4567-e89b-12d3-a456-426655440001"
                  username: "Filippo"
        '404':
          description: The user does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generic_response"
              example:
                status: "User not found"
  /users/{user_id}/following:
    get:
      tags: ["followers"]
      summary: Gets following users
      description: Get the users that a user is following
      operationId: getUserFollowing
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          schema:
            $ref: "#/components/schemas/uid"
          required: true
          description: The user ID of the user to follow.
      responses:
        '200':
          description: Returns the user list
          content:
            application/json:
              schema:
                type: array
                minItems: 0
                maxItems: 100
                description: The list of users that the user is following.
                items:
                  $ref: "#/components/schemas/uid_name"
              example:
                - user_id: "123e4567-e89b-12d3-a456-426655440000"
                  username: "Maria"
                - user_id: "123e4567-e89b-12d3-a456-426655440001"
                  username: "Filippo"
        '404':
          description: The user does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generic_response"
              example:
                status: "User not found"
  /users/{user_id}/followers/{follower_uid}:
    parameters:
      - name: user_id
        in: path
        schema:
          $ref: "#/components/schemas/uid"
        required: true
        description: The ID of the user to follow.
      - name: follower_uid
        in: path
        schema:
          $ref: "#/components/schemas/uid"
        required: true
        description: The follower's user ID.
    put:
      tags: ["followers"]
      summary: Follows a user
      description: Starts following a user
      operationId: followUser
      security:
        - BearerAuth: []
      responses:
        '201':
          description: Follow user action successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generic_response"
              example:
                status: "Success"
        '403':
          description: The user has no permission perform this action.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generic_response"
              example:
                status: "Forbidden"
        '404':
          description: The resource does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generic_response"
              example:
                status: "Resource not found"
        '400': # todo: not sure if this is the right error code
          description: Trying to follow a user that does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generic_response"
              example:
                status: "User not found"
    delete:
      tags: ["followers"]
      summary: Unfollows a user
      description: Stops following a user
      operationId: unfollowUser
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Unfollow user action successful.
        '404':
          description: The user is not followed by follower_uid, or the user does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generic_response"
              example:
                status: "User not found"
...
  /users/{user_id}/photos/{photo_id}/likes:
    get:
      tags: ["likes"]
      summary: Gets photo likes
      description: Get the list of users liking a photo
      operationId: getPhotoLikes
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          schema:
            $ref: "#/components/schemas/uid"
          required: true
          description: The ID of the author of the photo.
        - name: photo_id
          in: path
          schema:
            $ref: "#/components/schemas/photo_id"
          required: true
          description: The ID of the photo.
        - name: limit
          in: query
          schema:
            $ref: "#/components/schemas/limit"
          description: The number of elements to show.
          required: false
        - name: start_index
          in: query
          schema:
            $ref: "#/components/schemas/start_index"
          description: The starting offset.
          required: false
      responses:
        '200':
          description: Returns the user list
          content:
            application/json:
              schema:
                type: array
                minItems: 0
                maxItems: 100
                description: An array of users liking the photo.
                items:
                  $ref: "#/components/schemas/uid_name"
              example:
                - user_id: "123e4567-e89b-12d3-a456-426655440000"
                  username: "Maria"
                - user_id: "123e4567-e89b-12d3-a456-426655440001"
                  username: "Filippo"
        '404':
          description: The user or the photo does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generic_response"
              example:
                status: "Resource not found"
  /users/{user_id}/photos/{photo_id}/likes/{liker_uid}:
    parameters:
      - name: user_id
        in: path
        schema:
          $ref: "#/components/schemas/uid"
        required: true
        description: The ID of the author of the photo.
      - name: liker_uid
        in: path
        schema:
          $ref: "#/components/schemas/uid"
        required: true
        description: The ID of the user who likes the photo.
      - name: photo_id
        in: path
        schema:
          $ref: "#/components/schemas/photo_id"
        required: true
        description: The ID of the photo.
    put:
      tags: ["likes"]
      summary: likes a Photo
      description: Adds a like to a photo.
      operationId: likePhoto
      security:
        - BearerAuth: []
      responses:
        '201':
          description: Like photo action successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generic_response"
              example:
                status: "Success"
        '204':
          description: The user already likes the photo.
        '403':
          description: The user has no permission to perform this action.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generic_response"
              example:
                status: "Forbidden"
        '404':
          description: Resource not found (or the author of the photo has banned the authorized user).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generic_response"
              example:
                status: "Resource not found"
        '400':
          description: Bad URI (maybe liker_uid is invalid).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generic_response"
              example:
                status: "Bad photo_id"
    delete:
      tags: ["likes"]
      summary: Unlikes a photo
      description: Removes a like from a photo
      operationId: unlikePhoto
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Unlike photo action successful.
        '404':
          description: The user or photo does not exists, or the user is not liking the photo.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generic_response"
              example:
                status: "Resource not found"
...
  /users/{user_id}:
    get:
      tags: ["profile"]
      summary: Returns user profile
      description: Returns the profile of a user, including user's photos, followers, and following users.
      operationId: getUserProfile
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          schema:
            $ref: "#/components/schemas/uid"
          description: The user ID of the user to get the profile of.
          required: true
      responses:
        '200':
          description: Returns the profile details of the given user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/user_profile"
              example:
                username: "Maria"
                followers: 25
                following: 32
                followed: true
                photos: 50
        '404':
          description: User not found (or the authorized user is banned).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generic_response"
              example:
                status: "Resource not found"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    start_index:
      type: integer
      description: The starting offset.
      default: 0
    limit:
      type: integer
      description: The number of elements to show.
      default: 15
    uid_name:
      type: object
      properties:
        user_id:
          $ref: "#/components/schemas/uid"
        name:
          $ref: "#/components/schemas/name"
...
    uid:
      type: string
      minLength: 36
      maxLength: 36
      description: The user ID.
      format: uuid
      example: "1b4e28ba-2fa1-11d2-883f-0016d3cca427"
...
    name:
      type: string
      description: The username of the user.
      example: Maria
      pattern: '^[a-zA-Z0-9_.]*$'
      minLength: 3
      maxLength: 16
    photo_id:
      type: integer
      description: The ID of the photo.
      example: 1527
...
    followers_n:
      type: integer
      example: 420
      description: Number of followers.
    following_n:
      type: integer
      example: 69
      description: Number of following users.
...
    follow_status:
      type: boolean
      example: false
      description: Whether the user follows the other user.
    photos_n:
      type: integer
      example: 90
      description: Number of photos.
    user_profile:
      type: object
      description: The profile of the user.
      properties:
        username:
          $ref: "#/components/schemas/name"
        followers:
          $ref: "#/components/schemas/followers_n"
        following:
          $ref: "#/components/schemas/following_n"
        photos:
          $ref: "#/components/schemas/photos_n"
        followed:
          $ref: "#/components/schemas/follow_status"
...
    generic_response:
      type: object
      description: An object representing a success.
      properties:
        status:
          type: string
          pattern: ".*"
          description: The status of the request.
          example: "Success"
...